% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{msg_look_app}
\alias{msg_look_app}
\alias{msg_look_gadget}
\title{Run Outlook Message Looker App}
\usage{
msg_look_app(...)

msg_look_gadget(..., viewer = shiny::paneViewer())
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}} or \code{\link[shiny:runGadget]{shiny::runGadget()}}.}

\item{viewer}{Specify where the gadget should be displayed--viewer pane,
dialog window, or external browser--by passing in a call to one of the
\code{\link[=viewer]{viewer()}} functions.}
}
\description{
Runs a Shiny app or gadget to read an Outlook message by dragging the \code{.msg}
file into the window. The \code{.msg} file is converted to a self-contained
HTML document (including attachments) using the \pkg{msgxtractor} and
\pkg{base64enc} packages. You can also download the self-contained HTML file.
}
\section{Functions}{
\itemize{
\item \code{msg_look_app}: Run the msglooker app as a Shiny app.

\item \code{msg_look_gadget}: Run the msglooker app as an RStudio Gadget
}}

