% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{read_msg}
\alias{read_msg}
\alias{msg2html}
\title{Read and Write an Outlook Message to Self-Contained HTML}
\usage{
read_msg(path = file.choose())

msg2html(path = file.choose(), output = file.choose(new = TRUE))
}
\arguments{
\item{path}{The path to the msg}

\item{output}{The path to the new HTML file where the email will be saved}
}
\description{
Read an email saved from Outlook on Windows to a \code{.msg} file into R, or write
an email into a self-contained HTML file.
}
